separate univariate, multivariate and process distributions
