+++ inst/include/rmolib/random/multivariate/arnold_mo_distribution.hpp
@@ -1,13 +1,15 @@
 #pragma once
 
+#include <algorithm>
 #include <cmath>
 #include <limits>
 #include <type_traits>
 #include <vector>
 
-#include "rmolib/random/exponential_distribution.hpp"
-#include "rmolib/random/internal/is_within.hpp"
-#include "rmolib/random/internal/mo_param_type.hpp"
+#include "rmolib/random/multivariate/internal/is_within.hpp"
+#include "rmolib/random/multivariate/internal/mo_param_type.hpp"
+#include "rmolib/random/univariate/exponential_distribution.hpp"
+#include "rmolib/random/univariate/r_discrete_distribution.hpp"
 
 namespace rmolib {
 
@@ -15,8 +17,10 @@ namespace random {
 
 template <typename _Container,
           typename _UnitExponentialDistribution =
-              exponential_distribution<typename _Container::value_type>>
-class esm_mo_distribution {
+              exponential_distribution<typename _Container::value_type>,
+          typename _DiscreteDistribution = r_discrete_distribution<
+              typename _Container::size_type, typename _Container::value_type>>
+class arnold_mo_distribution {
  public:
   using result_type = _Container;
 
@@ -25,21 +29,21 @@ class esm_mo_distribution {
 
   class param_type {
    public:
-    using distribution_type = esm_mo_distribution;
+    using distribution_type = arnold_mo_distribution;
 
     param_type() = default;
 
-    template <typename _InputIterator>
-    explicit param_type(size_type dim, _InputIterator first,
-                        _InputIterator last)
-        : dim_{dim}, intensities_{first, last} {
+    template <typename _ForwardIterator>
+    explicit param_type(size_type dim, _ForwardIterator first,
+                        _ForwardIterator last)
+        : dim_{dim},
+          total_intensity_{std::accumulate(first, last, value_type{0})},
+          discrete_parm_{first, last} {
       __validate_input();
     }
 
     explicit param_type(size_type dim, _Container intensities)
-        : dim_{dim}, intensities_{intensities} {
-      __validate_input();
-    }
+        : param_type{dim, intensities.begin(), intensities.end()} {}
 
     param_type(size_type dim, std::initializer_list<value_type> wl)
         : param_type{dim, wl.begin(), wl.end()} {}
@@ -56,13 +60,20 @@ class esm_mo_distribution {
     // compiler generated ctor and assignment op is sufficient
 
     auto dim() const { return dim_; }
+    auto intensities() const {
+      auto out = discrete_parm_.probabilities();
+      std::transform(out.cbegin(), out.cend(), out.begin(),
+                     std::bind(std::multiplies<value_type>(),
+                               std::placeholders::_1, total_intensity_));
+      return out;
+    }
 
-    auto intensities() const { return intensities_; }
-
-    friend class esm_mo_distribution;
+    friend class arnold_mo_distribution;
 
     friend bool operator==(const param_type& lhs, const param_type& rhs) {
-      return lhs.dim_ == rhs.dim_ && lhs.intensities_ == rhs.intensities_;
+      return lhs.dim_ == rhs.dim_ &&
+             lhs.total_intensity_ == rhs.total_intensity_ &&
+             lhs.discrete_parm_ == rhs.discrete_parm_;
     }
 
     friend bool operator!=(const param_type& lhs, const param_type& rhs) {
@@ -71,42 +82,40 @@ class esm_mo_distribution {
 
    private:
     size_type dim_{1};
-    _Container intensities_{{value_type{1}}};
+    value_type total_intensity_{1};
+    typename _DiscreteDistribution::param_type discrete_parm_{{value_type{1}}};
 
     void __validate_input() const {
-      if (!((size_type{1} << dim_) == intensities_.size() + size_type{1}))
-        throw std::domain_error("intensities vector has wrongsize");
-
-      if (!std::all_of(intensities_.begin(), intensities_.end(),
-                       [](const auto v) { return v >= 0; }))
-        throw std::domain_error("negative intensities not allowed");
+      if (!((size_type{1} << dim_) ==
+            discrete_parm_.probabilities().size() + size_type{1}))
+        throw std::domain_error("intensities vector has wrong size");
     }
 
     static_assert(
         std::is_floating_point<value_type>::value,
-        "Class template rmolib::random::esm_mo_distribution<> must be "
+        "Class template rmolib::random::arnold_mo_distribution<> must be "
         "parametrized with floating point type");
   };
 
-  esm_mo_distribution() = default;
+  arnold_mo_distribution() = default;
 
-  template <typename _InputIterator>
-  explicit esm_mo_distribution(size_type dim, _InputIterator first,
-                               _InputIterator last)
+  template <typename _ForwardIterator>
+  explicit arnold_mo_distribution(size_type dim, _ForwardIterator first,
+                                  _ForwardIterator last)
       : parm_{dim, first, last} {
     init_unit_exponential_distribution();
   }
 
-  explicit esm_mo_distribution(size_type dim, _Container intensities)
+  explicit arnold_mo_distribution(size_type dim, _Container intensities)
       : parm_{dim, intensities} {}
 
-  explicit esm_mo_distribution(size_type dim,
-                               std::initializer_list<value_type>& wl)
+  explicit arnold_mo_distribution(size_type dim,
+                                  std::initializer_list<value_type>& wl)
       : parm_{dim, wl.begin(), wl.end()} {
     init_unit_exponential_distribution();
   }
 
-  explicit esm_mo_distribution(const param_type& parm) : parm_{parm} {
+  explicit arnold_mo_distribution(const param_type& parm) : parm_{parm} {
     init_unit_exponential_distribution();
   }
 
@@ -147,33 +156,38 @@ class esm_mo_distribution {
     // TODO: check compatibility
     using size_type = typename _Container::size_type;
 
-    std::fill(out.begin(), out.end(),
-              std::numeric_limits<value_type>::infinity());
-    for (size_type j = 0, num_shocks = parm.intensities_.size(); j < num_shocks;
-         ++j) {
-      if (parm.intensities_[j] > 0) {
-        auto shock_time =
-            unit_exponential_distribution_(engine) / parm.intensities_[j];
-        for (size_type i = 0; i < parm.dim_; ++i) {
-          if (internal::is_within(i, j)) out[i] = std::min(out[i], shock_time);
+    std::fill(out.begin(), out.end(), value_type{0});
+    std::vector<bool> destroyed(parm.dim_, false);
+
+    while (!std::all_of(destroyed.begin(), destroyed.end(),
+                        [](bool v) { return v; })) {
+      auto waiting_time =
+          unit_exponential_distribution_(engine) / parm.total_intensity_;
+      auto affected = discrete_distribution_(engine, parm.discrete_parm_);
+
+      for (size_type i = 0; i < parm.dim_; ++i) {
+        if (!destroyed[i]) {
+          out[i] += waiting_time;
+          if (internal::is_within(i, affected)) destroyed[i] = true;
         }
       }
     }
   }
 
-  friend bool operator==(const esm_mo_distribution& lhs,
-                         const esm_mo_distribution& rhs) {
+  friend bool operator==(const arnold_mo_distribution& lhs,
+                         const arnold_mo_distribution& rhs) {
     return lhs.parm_ == rhs.parm_;
   }
 
-  friend bool operator!=(const esm_mo_distribution& lhs,
-                         const esm_mo_distribution& rhs) {
+  friend bool operator!=(const arnold_mo_distribution& lhs,
+                         const arnold_mo_distribution& rhs) {
     return lhs.parm_ != rhs.parm_;
   }
 
  private:
   param_type parm_{};
   _UnitExponentialDistribution unit_exponential_distribution_{};
+  _DiscreteDistribution discrete_distribution_{};
 
   void init_unit_exponential_distribution() {
     if constexpr (std::is_constructible_v<_UnitExponentialDistribution,
@@ -187,9 +201,17 @@ class esm_mo_distribution {
   static_assert(
       std::is_same<value_type,
                    typename _UnitExponentialDistribution::result_type>::value,
-      "Class template rmolib::random::esm_mo_distribution<> must be "
+      "Class template rmolib::random::arnold_mo_distribution<> must be "
       "parametrized with unit_exponential_distribution-type with matching "
       "result_type");
+
+  // TODO: check static_assert
+  static_assert(
+      std::is_same<size_type,
+                   typename _DiscreteDistribution::result_type>::value,
+      "Class template rmolib::random::arnold_mo_distribution<> must be "
+      "parametrized with discrete_distribution-type with matching "
+      "size_type");
 };
 
 /*
@@ -198,13 +220,13 @@ class esm_mo_distribution {
   template <class _CharType, class _Traits, typename _Container, typename
   _UnitExponentialDistribution> std::basic_ostream<_CharType, _Traits>&
   operator<<(std::basic_ostream<_CharType, _Traits>& os,
-            esm_mo_distribution<_Container, _UnitExponentialDistribution>&
+            arnold_mo_distribution<_Container, _UnitExponentialDistribution>&
   dist);
 
   template <class _CharType, class _Traits, typename _Container, typename
   _UnitExponentialDistribution> std::basic_istream<_CharType, _Traits>&
   operator>>(std::basic_istream<_CharType, _Traits>& is,
-             esm_mo_distribution<_Container, _UnitExponentialDistribution>&
+             arnold_mo_distribution<_Container, _UnitExponentialDistribution>&
   dist);
 */
