+++ inst/include/rmolib/random/multivariate/cuadras_auge_distribution.hpp
@@ -4,7 +4,7 @@
 #include <limits>
 #include <type_traits>
 
-#include "rmolib/random/exponential_distribution.hpp"
+#include "rmolib/random/univariate/exponential_distribution.hpp"
 
 namespace rmolib {
