+++ inst/include/rmolib/random/univariate/deterministic_distribution.hpp
