+++ inst/include/rmolib/random/univariate/pareto_distribution.hpp
@@ -5,7 +5,7 @@
 #include <stdexcept>
 #include <type_traits>
 
-#include "rmolib/random/uniform_real_distribution.hpp"
+#include "rmolib/random/univariate/uniform_real_distribution.hpp"
 
 namespace rmolib {
 
@@ -28,7 +28,8 @@ struct is_pareto_param_type<
 template <typename _T>
 constexpr bool is_pareto_param_type_v = is_pareto_param_type<_T>::value;
 
-template <typename _RealType, typename _UnitUniformRealDistribution = uniform_real_distribution<_RealType>>
+template <typename _RealType, typename _UnitUniformRealDistribution =
+                                  uniform_real_distribution<_RealType>>
 class pareto_distribution {
  public:
   using result_type = _RealType;
