+++ src/test-arnold_mo_distribution.cpp
@@ -4,10 +4,10 @@
 
 #include <Rcpp.h>
 #include <r_engine.hpp>
-#include <rmolib/random/arnold_mo_distribution.hpp>
-#include <rmolib/random/exponential_distribution.hpp>
-#include <rmolib/random/r_discrete_distribution.hpp>
-#include <rmolib/random/uniform_real_distribution.hpp>
+#include <rmolib/random/multivariate/arnold_mo_distribution.hpp>
+#include <rmolib/random/univariate/exponential_distribution.hpp>
+#include <rmolib/random/univariate/r_discrete_distribution.hpp>
+#include <rmolib/random/univariate/uniform_real_distribution.hpp>
 #include <testthat.h>
 
 #include "testutils-approxequals.h"
