+++ src/test-pareto_distribution.cpp
@@ -2,8 +2,8 @@
 #include <cmath>
 
 #include <r_engine.hpp>
-#include <rmolib/random/pareto_distribution.hpp>
-#include <rmolib/random/uniform_real_distribution.hpp>
+#include <rmolib/random/univariate/pareto_distribution.hpp>
+#include <rmolib/random/univariate/uniform_real_distribution.hpp>
 #include <testthat.h>
 
 #define RMO_TEST_DIST_NAME pareto_distribution
